refactor(comb): Move iterator with other repeat combinators
