+++ src/combinator/multi.rs
@@ -9,6 +9,112 @@ use crate::stream::Stream;
 use crate::Parser;
 use crate::Result;
 
+/// Repeats the embedded parser, lazily returning the results
+///
+/// Call the iterator's [`ParserIterator::finish`] method to get the remaining input if successful,
+/// or the error value if we encountered an error.
+///
+/// On [`ErrMode::Backtrack`][crate::error::ErrMode::Backtrack], iteration will stop. To instead chain an error up, see [`cut_err`][crate::combinator::cut_err].
+///
+/// # Example
+///
+/// ```rust
+/// # use winnow::prelude::*;
+/// use winnow::{combinator::iterator, ascii::alpha1, combinator::terminated};
+/// use std::collections::HashMap;
+///
+/// let data = "abc|defg|hijkl|mnopqr|123";
+/// let mut it = iterator(data, terminated(alpha1, "|"));
+///
+/// let parsed = it.map(|v| (v, v.len())).collect::<HashMap<_,_>>();
+/// let res: ModalResult<_> = it.finish();
+///
+/// assert_eq!(parsed, [("abc", 3usize), ("defg", 4), ("hijkl", 5), ("mnopqr", 6)].iter().cloned().collect());
+/// assert_eq!(res, Ok(("123", ())));
+/// ```
+pub fn iterator<Input, Output, Error, ParseNext>(
+    input: Input,
+    parser: ParseNext,
+) -> ParserIterator<ParseNext, Input, Output, Error>
+where
+    ParseNext: Parser<Input, Output, Error>,
+    Input: Stream,
+    Error: ParserError<Input>,
+{
+    ParserIterator {
+        parser,
+        input,
+        state: State::Running,
+        o: Default::default(),
+    }
+}
+
+/// Main structure associated to [`iterator`].
+pub struct ParserIterator<F, I, O, E>
+where
+    F: Parser<I, O, E>,
+    I: Stream,
+{
+    parser: F,
+    input: I,
+    state: State<E>,
+    o: core::marker::PhantomData<O>,
+}
+
+impl<F, I, O, E> ParserIterator<F, I, O, E>
+where
+    F: Parser<I, O, E>,
+    I: Stream,
+    E: ParserError<I>,
+{
+    /// Returns the remaining input if parsing was successful, or the error if we encountered an error.
+    pub fn finish(self) -> Result<(I, ()), E> {
+        match self.state {
+            State::Running | State::Done => Ok((self.input, ())),
+            State::Cut(e) => Err(e),
+        }
+    }
+}
+
+impl<F, I, O, E> core::iter::Iterator for &mut ParserIterator<F, I, O, E>
+where
+    F: Parser<I, O, E>,
+    I: Stream,
+    E: ParserError<I>,
+{
+    type Item = O;
+
+    fn next(&mut self) -> Option<Self::Item> {
+        if matches!(self.state, State::Running) {
+            let start = self.input.checkpoint();
+
+            match self.parser.parse_next(&mut self.input) {
+                Ok(o) => {
+                    self.state = State::Running;
+                    Some(o)
+                }
+                Err(e) if e.is_backtrack() => {
+                    self.input.reset(&start);
+                    self.state = State::Done;
+                    None
+                }
+                Err(e) => {
+                    self.state = State::Cut(e);
+                    None
+                }
+            }
+        } else {
+            None
+        }
+    }
+}
+
+enum State<E> {
+    Running,
+    Done,
+    Cut(E),
+}
+
 /// [`Accumulate`] the output of a parser into a container, like `Vec`
 ///
 /// This stops before `n` when the parser returns [`ErrMode::Backtrack`][crate::error::ErrMode::Backtrack]. To instead chain an error up, see
